requirements.txt
torch>=1.1.0
torchvision
numpy
pillow